refactor(server): segregate api tests
